--- tools/gfx/cpu/render-cpu.h
@@ -1,11 +1,0 @@
-// render-cpu.h
-#pragma once
-
-#include "../renderer-shared.h"
-
-namespace gfx
-{
-
-SlangResult SLANG_MCALL createCPUDevice(const IDevice::Desc* desc, IDevice** outDevice);
-
-}
